# Improve performance of Network.copy