In a copy of a `Network` new `s` parameters are assigned. This triggers a call to `__generate_secondary_properties` which is not required as `__generate_secondary_properties` depends only on class properties, not on instance properties (or the values of the s parameters)

On this data
```
import numpy as np
import skrf as rf

f = rf.Frequency.from_f(np.linspace(2e6, 3e6, 10), unit="Hz")
self=n=rf.Network(frequency=f, s=np.random.rand(10,)*1j, name='test')
``` 
a benchmark `%timeit n.copy()` is improved by:
```
Master: 1.06 ms ± 143 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
PR: 217 µs ± 32.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```